18.06.
Basic Enemy State Machine, Very Easy Enemy Type, Ini Check

+++ Assets/MonsterLinker/Scripts/ScriptableObjects/Attack.cs
@@ -12,5 +12,5 @@ public class Attack : ScriptableObject
     [Tooltip("Health recovered by using this attack: % of max HP value")]
     public float HPGain;
     [Tooltip("Speed of the attack")]
-    public float Speed;    
+    public int Speed;    
 }
